Added ATD_AIRFLOW_PATH

+++ script/entrypoint.sh
@@ -1,57 +1,44 @@
 #!/usr/bin/env bash
 
-AIRFLOW_IGNORE_PATH="/usr/local/airflow/dags/.airflowignore";
+# Create an alias for AIRFLOW_HOME
+export ATD_AIRFLOW_PATH="/usr/local/airflow";
+
+AIRFLOW_IGNORE_PATH="${ATD_AIRFLOW_PATH}/.airflowignore";
 # Be sure to always remove it before starting up...
-rm $AIRFLOW_IGNORE_PATH;
+if [[ -f "${AIRFLOW_IGNORE_PATH}" ]]
+then
+    echo "Removing existing airflow ignore file...";
+    rm $AIRFLOW_IGNORE_PATH;
+fi
+
 
 if [[ "${ATD_AIRFLOW_ENVIRONMENT}" = "production" ]] ; then
-    cp /usr/local/airflow/config/.airflowignore $AIRFLOW_IGNORE_PATH;
+    cp $ATD_AIRFLOW_PATH/config/.airflowignore $AIRFLOW_IGNORE_PATH;
 else
-    cp /usr/local/airflow/config/.airflowignore-dev $AIRFLOW_IGNORE_PATH;
+    cp $ATD_AIRFLOW_PATH/config/.airflowignore-dev $AIRFLOW_IGNORE_PATH;
 fi;
 
 TRY_LOOP="20"
 
-: "${REDIS_HOST:="redis"}"
-: "${REDIS_PORT:="6379"}"
-: "${REDIS_PASSWORD:=""}"
-
-: "${POSTGRES_HOST:="postgres"}"
-: "${POSTGRES_PORT:="5432"}"
-: "${POSTGRES_USER:="airflow"}"
-: "${POSTGRES_PASSWORD:="airflow"}"
-: "${POSTGRES_DB:="airflow"}"
-
-# Defaults and back-compat
+# Global defaults and back-compat
 : "${AIRFLOW_HOME:="/usr/local/airflow"}"
 : "${AIRFLOW__CORE__FERNET_KEY:=${FERNET_KEY:=$(python -c "from cryptography.fernet import Fernet; FERNET_KEY = Fernet.generate_key().decode(); print(FERNET_KEY)")}}"
 : "${AIRFLOW__CORE__EXECUTOR:=${EXECUTOR:-Sequential}Executor}"
 
+# Load DAGs examples (default: Yes)
+if [[ -z "$AIRFLOW__CORE__LOAD_EXAMPLES" && "${LOAD_EX:=n}" == n ]]; then
+  AIRFLOW__CORE__LOAD_EXAMPLES=False
+fi
+
 export \
   AIRFLOW_HOME \
-  AIRFLOW__CELERY__BROKER_URL \
-  AIRFLOW__CELERY__RESULT_BACKEND \
   AIRFLOW__CORE__EXECUTOR \
   AIRFLOW__CORE__FERNET_KEY \
   AIRFLOW__CORE__LOAD_EXAMPLES \
-  AIRFLOW__CORE__SQL_ALCHEMY_CONN \
-
-
-# Load DAGs exemples (default: Yes)
-if [[ -z "$AIRFLOW__CORE__LOAD_EXAMPLES" && "${LOAD_EX:=n}" == n ]]
-then
-  AIRFLOW__CORE__LOAD_EXAMPLES=False
-fi
 
 # Install custom python package if requirements.txt is present
 if [ -e "/requirements.txt" ]; then
     $(command -v pip) install --user -r /requirements.txt
-fi
-
-if [ -n "$REDIS_PASSWORD" ]; then
-    REDIS_PREFIX=:${REDIS_PASSWORD}@
-else
-    REDIS_PREFIX=
 fi
 
 wait_for_port() {
@@ -68,14 +55,68 @@ wait_for_port() {
   done
 }
 
+# Other executors than SequentialExecutor drive the need for an SQL database, here PostgreSQL is used
 if [ "$AIRFLOW__CORE__EXECUTOR" != "SequentialExecutor" ]; then
-  AIRFLOW__CORE__SQL_ALCHEMY_CONN="postgresql+psycopg2://$POSTGRES_USER:$POSTGRES_PASSWORD@$POSTGRES_HOST:$POSTGRES_PORT/$POSTGRES_DB"
-  AIRFLOW__CELERY__RESULT_BACKEND="db+postgresql://$POSTGRES_USER:$POSTGRES_PASSWORD@$POSTGRES_HOST:$POSTGRES_PORT/$POSTGRES_DB"
+  # Check if the user has provided explicit Airflow configuration concerning the database
+  if [ -z "$AIRFLOW__CORE__SQL_ALCHEMY_CONN" ]; then
+    # Default values corresponding to the default compose files
+    : "${POSTGRES_HOST:="postgres"}"
+    : "${POSTGRES_PORT:="5432"}"
+    : "${POSTGRES_USER:="airflow"}"
+    : "${POSTGRES_PASSWORD:="airflow"}"
+    : "${POSTGRES_DB:="airflow"}"
+    : "${POSTGRES_EXTRAS:-""}"
+
+    AIRFLOW__CORE__SQL_ALCHEMY_CONN="postgresql+psycopg2://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}${POSTGRES_EXTRAS}"
+    export AIRFLOW__CORE__SQL_ALCHEMY_CONN
+
+    # Check if the user has provided explicit Airflow configuration for the broker's connection to the database
+    if [ "$AIRFLOW__CORE__EXECUTOR" = "CeleryExecutor" ]; then
+      AIRFLOW__CELERY__RESULT_BACKEND="db+postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}${POSTGRES_EXTRAS}"
+      export AIRFLOW__CELERY__RESULT_BACKEND
+    fi
+  else
+    if [[ "$AIRFLOW__CORE__EXECUTOR" == "CeleryExecutor" && -z "$AIRFLOW__CELERY__RESULT_BACKEND" ]]; then
+      >&2 printf '%s\n' "FATAL: if you set AIRFLOW__CORE__SQL_ALCHEMY_CONN manually with CeleryExecutor you must also set AIRFLOW__CELERY__RESULT_BACKEND"
+      exit 1
+    fi
+
+    # Derive useful variables from the AIRFLOW__ variables provided explicitly by the user
+    POSTGRES_ENDPOINT=$(echo -n "$AIRFLOW__CORE__SQL_ALCHEMY_CONN" | cut -d '/' -f3 | sed -e 's,.*@,,')
+    POSTGRES_HOST=$(echo -n "$POSTGRES_ENDPOINT" | cut -d ':' -f1)
+    POSTGRES_PORT=$(echo -n "$POSTGRES_ENDPOINT" | cut -d ':' -f2)
+  fi
+
   wait_for_port "Postgres" "$POSTGRES_HOST" "$POSTGRES_PORT"
 fi
 
+# CeleryExecutor drives the need for a Celery broker, here Redis is used
 if [ "$AIRFLOW__CORE__EXECUTOR" = "CeleryExecutor" ]; then
-  AIRFLOW__CELERY__BROKER_URL="redis://$REDIS_PREFIX$REDIS_HOST:$REDIS_PORT/1"
+  # Check if the user has provided explicit Airflow configuration concerning the broker
+  if [ -z "$AIRFLOW__CELERY__BROKER_URL" ]; then
+    # Default values corresponding to the default compose files
+    : "${REDIS_PROTO:="redis://"}"
+    : "${REDIS_HOST:="redis"}"
+    : "${REDIS_PORT:="6379"}"
+    : "${REDIS_PASSWORD:=""}"
+    : "${REDIS_DBNUM:="1"}"
+
+    # When Redis is secured by basic auth, it does not handle the username part of basic auth, only a token
+    if [ -n "$REDIS_PASSWORD" ]; then
+      REDIS_PREFIX=":${REDIS_PASSWORD}@"
+    else
+      REDIS_PREFIX=
+    fi
+
+    AIRFLOW__CELERY__BROKER_URL="${REDIS_PROTO}${REDIS_PREFIX}${REDIS_HOST}:${REDIS_PORT}/${REDIS_DBNUM}"
+    export AIRFLOW__CELERY__BROKER_URL
+  else
+    # Derive useful variables from the AIRFLOW__ variables provided explicitly by the user
+    REDIS_ENDPOINT=$(echo -n "$AIRFLOW__CELERY__BROKER_URL" | cut -d '/' -f3 | sed -e 's,.*@,,')
+    REDIS_HOST=$(echo -n "$POSTGRES_ENDPOINT" | cut -d ':' -f1)
+    REDIS_PORT=$(echo -n "$POSTGRES_ENDPOINT" | cut -d ':' -f2)
+  fi
+
   wait_for_port "Redis" "$REDIS_HOST" "$REDIS_PORT"
 fi
 
@@ -89,7 +130,7 @@ case "$1" in
     exec airflow webserver
     ;;
   worker|scheduler)
-    # To give the webserver time to run initdb.
+    # Give the webserver time to run initdb.
     sleep 10
     exec airflow "$@"
     ;;
